O'zgartirishlar kiritildi

+++ src/Begins/begin4.java
@@ -1,12 +1,10 @@
 package Begins;
-
 import java.util.Scanner;
-
 public class begin4 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.print("Enter d:");
-      int d = sc.nextInt();
+      double d = sc.nextDouble();
       double pi=Math.PI;
       double Length=pi*d;
       System.out.println("Length = "+Length);
